CDAP-19078 update service account loading to use newer libraries
Configure the GCS connector to use a custom AccessTokenProvider
that uses the newer Credentials library instead of the default
behavior, which uses the Credential library. This allows the
connector to work with newer service account types, such as the
external_account type.

As part of this, refactoring the util class to clean up the code
and also remove bad assumptions, like the service account will
always be of type 'service_account' and instead let the
GoogleCredentials class look at the type and determine the correct
underlying implementation.

Also removing a bunch of useless constants and property prefixing.